requirements.txt
ImageHash==3.4
matplotlib==2.0.2
scipy==0.19.1
Pillow==5.0.0
